Move ServerConfig class to sub-package
